Separate processors from Config.
